Put the C++ classes into cpp/hpp files. Kept the C-API around for posterity's sake. Moved the networking common include to its own header file.
