+++ demos/client_server/client.cpp
@@ -1,4 +1,4 @@
-#include "junkworks/basic_udp.hpp"
+#include "junkworks/udp_socket.hpp"
 
 #include <iostream>
