+++ junkworks/CMakeLists.txt
@@ -1,11 +1,14 @@
 set(
    junkwork_sources
    src/basic_udp.cpp
+   src/udp_socket.cpp
 )
 
 set(
    junkwork_headers
    include/junkworks/basic_udp.hpp
+   include/junkworks/network_common.h
+   include/junkworks/udp_socket.hpp
 )
 
 set(
